feat: configure initial packages and architecture
The purpose of this project is be a start of others project, so
the first thing that I did was configure all essentials tools to this.
Basicaly to all projects we need a well configured Framework (more than
a single file with all responsability) and a ORM to improve database
comunication. To do this things I setup a configuration of Flask using
blueprints to facilitate controllers separation, also, was created a
basic configuration of SQLAlchemy. As support tools was added Migrations
to handle changes in db structure and Flask-Script to help with server
start and db migration operation.